moved path roots in fog05.interfaces.Constrants with a commodity function to concatenate paths
