+++ fog05/interfaces/Constants.py
@@ -1,0 +1,8 @@
+
+aroot = '/afos/'
+droot = '/dfos/'
+sroot = '/sfos/'
+
+
+def append_to_path(base, extention):
+    return '{}/{}'.format(base, extention)
