+++ fog05/store.py
@@ -5,6 +5,50 @@ from yaks import Selector
 from yaks import Value
 
 
+class FOSStore(object):
+    """
+    Helper class to interact with the Store
+    """
+
+    def __init__(self, server, aroot, droot, sroot, home):
+        '''
+
+        Initialize the Store with root and home
+
+        :param server: the address of the YAKS server
+        :param aroot: actual store root
+        :param droot: desired store root
+        :param home: store home also used to generate store id
+        '''
+
+        self.y = YAKS()
+        self.y.login(server)
+
+        self.aroot = aroot  # '/dfos/{}'
+        self.ahome = '{}/{}'.format(aroot, home)
+
+        self.droot = droot  # '/dfos/{}'
+        self.dhome = '{}/{}'.format(droot, home)
+
+        self.sroot = sroot
+        self.shome = '{}/{}'.format(sroot, home)
+
+        self.actual = Store(self.y, self.aroot, self.ahome, 1024)
+        self.desired = Store(self.y, self.droot, self.dhome, 1024)
+        self.system = Store(self.y,  self.sroot, self.shome, 1024)
+
+    def close(self):
+        '''
+        Close the store
+
+        :return: None
+        '''
+        self.actual.close()
+        self.desired.close()
+        self.system.close()
+        self.y.logout()
+
+
 class Store(object):
 
     def __init__(self, api, root_path, home_path, cachesize):
